requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.16,<4.0
django-filter>=25.1
djangorestframework-simplejwt>=5.4
django-cors-headers>=4.7
python-decouple>=3.8
Pillow>=11.1
